+++ cli/cli/commands/enclave/rm/rm.go
@@ -9,6 +9,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/kurtosis_command"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/kurtosis_command/args"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/kurtosis_command/flags"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/command_framework/prebuilt_command_components/enclave_id_arg"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/defaults"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/helpers/engine_manager"
@@ -22,8 +23,14 @@ import (
 )
 
 const (
-	shouldForceRemoveFlagKey = "force"
 	enclaveIdArgKey          = "enclave-id"
+	shouldForceRemoveFlagKey = "force"
+
+	engineClientCtxKey = "engine-client"
+	engineClientCloseFuncCtxKey = "engine-client-close-func"
+
+	isEnclaveIdArgOptional = false
+	isEnclaveIdArgGreedy = true
 
 	defaultShouldForceRemove = "false"
 )
@@ -41,20 +48,53 @@ var EnclaveRmCmd = &kurtosis_command.KurtosisCommand{
 			Default:   defaultShouldForceRemove,
 		},
 	},
-	// TODO Use a prebuilt enclaveIdArg component here!!!
 	Args:             []*args.ArgConfig{
-		{
-			Key:             enclaveIdArgKey,
-			IsGreedy:        true,
-			CompletionsFunc: nil,
-			ValidationFunc:  nil,
-		},
+		enclave_id_arg.NewEnclaveIDArg(
+			enclaveIdArgKey,
+			engineClientCtxKey,
+			isEnclaveIdArgOptional,
+			isEnclaveIdArgGreedy,
+		),
 	},
+	PreValidationAndRunFunc: setup,
 	RunFunc:          run,
+	PostValidationAndRunFunc: teardown,
 }
 
-func run(flags *flags.ParsedFlags, args *args.ParsedArgs) error {
-	ctx := context.Background()
+// TODO generalize this to something like EngineUsingKurtosisCommand or something
+func setup(ctx context.Context) (context.Context, error) {
+	result := ctx
+
+	dockerClient, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred creating the Docker client")
+	}
+	dockerManager := docker_manager.NewDockerManager(
+		logrus.StandardLogger(),
+		dockerClient,
+	)
+
+	engineManager := engine_manager.NewEngineManager(dockerManager)
+	objAttrsProvider := schema.GetObjectAttributesProvider()
+	engineClient, closeClientFunc, err := engineManager.StartEngineIdempotentlyWithDefaultVersion(ctx, objAttrsProvider, defaults.DefaultEngineLogLevel)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred creating a new Kurtosis engine client")
+	}
+	result = context.WithValue(result, engineClientCtxKey, engineClient)
+	result = context.WithValue(result, engineClientCloseFuncCtxKey, closeClientFunc)
+
+	return result, nil
+}
+
+func run(ctx context.Context, flags *flags.ParsedFlags, args *args.ParsedArgs) error {
+	uncastedEngineClient := ctx.Value(engineClientCtxKey)
+	if uncastedEngineClient == nil {
+		return stacktrace.NewError("Expected an engine client to have been stored in the context under key '%v', but none was found; this is a bug in Kurtosis!", engineClientCtxKey)
+	}
+	engineClient, ok := uncastedEngineClient.(kurtosis_engine_rpc_api_bindings.EngineServiceClient)
+	if !ok {
+		return stacktrace.NewError("Found an object that should be the engine client stored in the context under key '%v', but this object wasn't of the correct type", engineClientCtxKey)
+	}
 
 	inputtedEnclaveIds, err := args.GetGreedyArg(enclaveIdArgKey)
 	if err != nil {
@@ -73,22 +113,6 @@ func run(flags *flags.ParsedFlags, args *args.ParsedArgs) error {
 	logrus.Debugf("Unique enclave IDs to destroy: %+v", enclaveIdsToDestroy)
 
 	logrus.Info("Destroying enclaves...")
-	dockerClient, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred creating the Docker client")
-	}
-	dockerManager := docker_manager.NewDockerManager(
-		logrus.StandardLogger(),
-		dockerClient,
-	)
-	engineManager := engine_manager.NewEngineManager(dockerManager)
-	objAttrsProvider := schema.GetObjectAttributesProvider()
-	engineClient, closeClientFunc, err := engineManager.StartEngineIdempotentlyWithDefaultVersion(ctx, objAttrsProvider, defaults.DefaultEngineLogLevel)
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred creating a new Kurtosis engine client")
-	}
-	defer closeClientFunc()
-
 	getEnclavesResp, err := engineClient.GetEnclaves(ctx, &emptypb.Empty{})
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred getting enclaves to check that the ones to destroy are stopped")
@@ -120,6 +144,25 @@ func run(flags *flags.ParsedFlags, args *args.ParsedArgs) error {
 	logrus.Info("Enclaves successfully destroyed")
 
 	return nil
+}
+
+func teardown(ctx context.Context) {
+	uncastedEngineClientCloseFunc := ctx.Value(engineClientCloseFuncCtxKey)
+	if uncastedEngineClientCloseFunc != nil {
+		engineClientCloseFunc, ok := uncastedEngineClientCloseFunc.(func() error)
+		if ok {
+			if err := engineClientCloseFunc(); err != nil {
+				logrus.Warnf("We tried to close the engine client after we're done using it, but doing so threw an error:\n%v", err)
+			}
+		} else {
+			logrus.Errorf("Expected the object at context key '%v' to be an engine client close function, but it wasn't; this is a bug in Kurtosis!", engineClientCloseFuncCtxKey)
+		}
+	} else {
+		logrus.Errorf(
+			"Expected to find an engine client close function during teardown at context key '%v', but none was found; this is a bug in Kurtosis!",
+			engineClientCloseFuncCtxKey,
+		)
+	}
 }
 
 // ====================================================================================================