+++ cli/docs/changelog.md
@@ -4,11 +4,13 @@
 * Added a "Debugging User Issues" section to the README
 * Added a `KurtosisCommand` wrapper over the `cobra.Command` objects that we're using to create CLI commands, so that we have a centralized place to add autocompletion
 * Added flags to `KurtosisCommand`
+* Added a `NewEnclaveIDArg` generator for building enclave ID args with tab-completion and validation out-of-the-box
 
 ### Changes
 * Switched the `enclave inspect` to the new command framework
 * Cleaned up our Kurtosis-custom logic wrapping the Cobra commands
 * Switched `enclave rm` to use the new command framework
+* Switched `enclave rm` and `enclave inspect` to use the `NewEnclaveIDArg` component
 
 ### Fixes
 * Fix the Kurtosis completion